This series addresses some critical bugs.

0. the number of max bytes to encode a varint must be 10 w/ our decoder
1. we fix the timestamp-delta overflow which must be in 64-bits
2. we fix varint shift overflow by adopting upstream protobuf c++ imlpementation
  for encoding and leveldb implementation for decoding varint

Testing: all unit tests + librdkafka test 0054-offset-timestamp no longer crashes
        redpanda.

TESTS=0054 make

Note that librdkafka test 0054 now fails due to a logical bug in what I think
is at the API level with an off-by-one error, unrelated to the crash of this series

* 'bugfix-ch934-timestamp-overflow-v1' of github.com:senior7515/v:
  storage:segment.cc:cosmetic:increase assert verbositoy on overflow
  varint: encoding MUST use 10 bytes
  utils:vint: port the protobuf varint implementation
  model: add missing fields to output operator for model::record
  kafka:batch_adapter: add error logging during batch parsing failure
  fundamental:record: use int64_t for timestamp_delta

+++ src/js/modules/rpc/parser.ts
@@ -173,7 +173,7 @@ export class Deserializer {
   record(bytes: Buffer) {
     const sizeBytes = bytes.readInt32LE(rcId.sizeBytes);
     const attributes = bytes.readInt8(rcId.attributes);
-    const timestampDelta = bytes.readInt32LE(rcId.timestampDelta);
+    const timestampDelta = bytes.readInt64LE(rcId.timestampDelta);
     const offsetDelta = bytes.readInt32LE(rcId.offsetDelta);
     const keySize = bytes.readInt32LE(rcId.keySize);
     const keyBytesToFollow = bytes.readInt32LE(rcId.keyBytesToFollow);
@@ -336,7 +336,7 @@ export class Serializer {
     let buf1: Buffer = Buffer.allocUnsafe(rcId.size);
     buf1.writeInt32LE(header.sizeBytes, rcId.sizeBytes);
     buf1.writeInt8(header.recordAttributes, rcId.attributes);
-    buf1.writeInt32LE(header.timestampDelta, rcId.timestampDelta);
+    buf1.writeInt64LE(header.timestampDelta, rcId.timestampDelta);
     buf1.writeInt32LE(header.offsetDelta, rcId.offsetDelta);
     buf1.writeInt32LE(header.keySize, rcId.keySize);
     buf1.writeInt32LE(header.key.length, rcId.keyBytesToFollow);
